ReactJS: Move BeanPool into a sub-package.
